+++ infrastructure/database/migrations/versions/a6b7c8d9e0f1_add_contract_tags.py
@@ -1,7 +1,7 @@
 """add tags to contracts"""
 
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 revision = 'a6b7c8d9e0f1'
 down_revision = 'f5a6b7c8d9e0'